requirements.txt
boto3==1.9.57
botocore==1.12.57
docutils==0.14
jmespath==0.9.3
Pillow==5.3.0
python-dateutil==2.7.5
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
